+++ oop-propro/src/test/java/ru/ssau/tk/oop/propro/MyArraysTest.java
@@ -270,9 +270,9 @@ public class MyArraysTest {
 
     @Test
     public void meanValueArray() {
-        int[] newFirstArray = {2, 2, 5, 7};
+        double[] newFirstArray = {2, 2, 5, 7};
         assertEquals(firstArray.meanValueArray(newFirstArray), 4, DELTA);
-        int[] newSecondArray = {-2, 5, -6};
+        double[] newSecondArray = {-2, 5, -6};
         assertEquals(firstArray.meanValueArray(newSecondArray), -1, DELTA);
     }
 }
